requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-jwt-extended==4.6.0
requests==2.26.0
python-dotenv==0.19.0
openai==0.27.0
bcrypt==3.2.0
werkzeug==2.0.3 